Rework leave balance to deduct from a single starting balance
Admins now set a "Starting Balance" that leave requests are actually
deducted from; the entitled-days quota is kept as an informational
reference only and no longer feeds into the available-balance math
(previously entitled + adjustment were summed, double-counting the
quota). Also removes the target-available-balance convenience field
added for HR imports, which proved more confusing than helpful.

+++ api/src/main/java/com/digileave/api/dto/LeaveSummaryDto.java
@@ -6,23 +6,26 @@ import lombok.Data;
 /**
  * Breakdown of a user's annual leave balance returned by GET /api/leave/summary/{userId}.
  *
- * available = entitled + transferred + startingBalanceAdjustment − used
+ * available = transferred + startingBalanceAdjustment − used
  */
 @Data
 @AllArgsConstructor
 public class LeaveSummaryDto {
 
-    /** Days awarded for the current calendar year. */
-    private int entitled;
+    /**
+     * Days awarded for the current calendar year (supports half-day entitlements).
+     * Informational only — shown to the user but excluded from {@code available}.
+     */
+    private double entitled;
 
     /** Unused days carried over from the previous year. */
     private int transferred;
 
     /**
-     * Manual one-off accounting adjustment (positive = credit, negative = deduction).
-     * Included so the frontend can show "includes ±N day adjustment" when non-zero.
+     * The starting balance leave requests are deducted from (set by an admin,
+     * e.g. from an HR export's "days left today").
      */
-    private int startingBalanceAdjustment;
+    private double startingBalanceAdjustment;
 
     /** Working days consumed by APPROVED annual-leave requests (supports 0.5 increments). */
     private double used;
@@ -30,6 +33,6 @@ public class LeaveSummaryDto {
     /** Working days in PENDING annual-leave requests (supports 0.5 increments). */
     private double pending;
 
-    /** Derived: entitled + transferred + startingBalanceAdjustment − used (supports 0.5 increments). */
+    /** Derived: transferred + startingBalanceAdjustment − used (supports 0.5 increments). */
     private double available;
 }
